Rename parsedString.js
