+++ validator/event/validator.js
@@ -1,6 +1,6 @@
 const utils = require('../../utils')
 const { ParsedHedTag } = require('../parser/types')
-const ParsedHedString = require('../parser/parsedString')
+const ParsedHedString = require('../parser/parsedHedString')
 const { generateIssue } = require('../../common/issues/issues')
 const { Schemas } = require('../../common/schema')
