+++ validator/parser/parsedHedString.js
